config: split IP/CIDR validators out of schema_validators.go
Move the interface/RA IP-address and CIDR leaf validators
(ValidateIPAddress, ValidateIPv6Address, pref64PrefixLengths,
ValidatePREF64CIDR, ValidateIPv4CIDR, ValidateIPv6CIDR, parseCIDRStrict)
into a new schema_validators_network.go sibling. parseCIDRStrict is the
shared CIDR parser also used by the routing validators (same package).

Refs #4421 (codex-173 #1): schema_validators.go is a cross-domain
validator grab-bag (1159 LOC, 47 top-level decls). Pure code motion --
declarations moved verbatim (doc comments + bodies unchanged); only the
per-file package/import headers are synthesized. The generic value
primitives (LeafValidator, ValidateEnum, ValidateInteger[Min],
ValidatePercent, duration/wire const bounds) stay in schema_validators.go.
