+++ pkg/config/schema_validators.go
@@ -143,142 +143,6 @@ const (
 	maxWireI32 = int64(math.MaxInt32)
 )
 
-// IP / CIDR validators (#1319 PR 3, interfaces subsystem). They reuse the
-// net package parsers the runtime consumers use (net.ParseIP /
-// net.ParseCIDR), so schema acceptance mirrors runtime parse acceptance
-// exactly — including the family classification rule ip.To4() != nil that
-// pkg/dataplane/userspace/interfaces.go buildConfiguredAddressSnapshots
-// applies to configured addresses.
-
-// ValidateIPAddress accepts any IPv4 or IPv6 address WITHOUT a prefix
-// length. Used for leaves the runtime feeds to net.ParseIP (e.g. GRE
-// tunnel source/destination, pkg/routing/tunnel.go:194), where garbage
-// silently disables the feature today.
-func ValidateIPAddress(raw string, _ *Config) error {
-	trimmed := strings.TrimSpace(raw)
-	if trimmed == "" {
-		return fmt.Errorf("missing value (expected an IP address, e.g. 10.0.1.10 or 2001:db8::1)")
-	}
-	if net.ParseIP(trimmed) == nil {
-		if _, _, err := net.ParseCIDR(trimmed); err == nil {
-			return fmt.Errorf("prefix length not allowed here (got %q; use a bare IP address)", raw)
-		}
-		return fmt.Errorf("not a valid IP address (got %q)", raw)
-	}
-	return nil
-}
-
-// ValidateIPv6Address accepts an IPv6 literal WITHOUT a prefix length and
-// rejects IPv4. Used for the RDNSS dns-server-address leaf (#2497): the RA
-// sender (pkg/ra/sender.go buildRA) feeds each address to netip.ParseAddr
-// and appends it to an RFC 8106 RecursiveDNSServer option, which is an
-// IPv6-only NDP option. The runtime skips an unparseable string but does
-// NOT family-gate, so a bare IPv4 literal (8.8.8.8) parses and is
-// advertised on the wire inside an IPv6 RDNSS option — a malformed RA.
-// The family rule mirrors the runtime's ip.To4()==nil classification.
-func ValidateIPv6Address(raw string, _ *Config) error {
-	trimmed := strings.TrimSpace(raw)
-	if trimmed == "" {
-		return fmt.Errorf("missing value (expected an IPv6 address, e.g. 2001:db8::1)")
-	}
-	ip := net.ParseIP(trimmed)
-	if ip == nil {
-		if _, _, err := net.ParseCIDR(trimmed); err == nil {
-			return fmt.Errorf("prefix length not allowed here (got %q; use a bare IPv6 address)", raw)
-		}
-		return fmt.Errorf("not a valid IP address (got %q)", raw)
-	}
-	if ip.To4() != nil {
-		return fmt.Errorf("not an IPv6 address (got %q; the RDNSS option is IPv6-only per RFC 8106)", raw)
-	}
-	return nil
-}
-
-// pref64PrefixLengths are the prefix lengths RFC 8781 §4 permits for a
-// PREF64 (NAT64 prefix) option: only these encode in the 3-bit PLC field.
-// A prefix of any other length cannot be advertised at all.
-var pref64PrefixLengths = map[int]struct{}{
-	32: {}, 40: {}, 48: {}, 56: {}, 64: {}, 96: {},
-}
-
-// ValidatePREF64CIDR accepts an IPv6 prefix usable in a PREF64 option:
-// a valid IPv6 CIDR (reusing ValidateIPv6CIDR's family + prefix rules)
-// whose prefix length is one of the RFC 8781 §4 set {32,40,48,56,64,96}.
-// Used for the router-advertisement nat-prefix/nat64prefix leaves (#2497):
-// the RA sender wraps the prefix in an ndp.PREF64 option whose wire PLC
-// field can only encode those six lengths. An out-of-set length committed
-// today reaches netip.ParsePrefix cleanly, so the runtime accepts it and
-// then either omits the option or mis-encodes the length.
-func ValidatePREF64CIDR(raw string, cfg *Config) error {
-	if err := ValidateIPv6CIDR(raw, cfg); err != nil {
-		return err
-	}
-	_, ipnet, err := net.ParseCIDR(strings.TrimSpace(raw))
-	if err != nil {
-		// Unreachable: ValidateIPv6CIDR already parsed it.
-		return fmt.Errorf("not a valid address/prefix-length (got %q): %v", raw, err)
-	}
-	ones, _ := ipnet.Mask.Size()
-	if _, ok := pref64PrefixLengths[ones]; !ok {
-		return fmt.Errorf("invalid PREF64 prefix length /%d (got %q; RFC 8781 permits only /32, /40, /48, /56, /64, /96)", ones, raw)
-	}
-	return nil
-}
-
-// ValidateIPv4CIDR accepts an IPv4 address with an explicit prefix
-// length (e.g. 10.0.1.10/24). The prefix is REQUIRED: every runtime
-// consumer of configured interface addresses net.ParseCIDRs the string
-// and silently skips it on error (dataplane snapshot:
-// pkg/dataplane/userspace/interfaces.go:391; RETH link-local checks:
-// pkg/daemon/daemon_reth.go:308), so a bare IP would commit and then
-// silently not exist. The v4/v6 family split mirrors the runtime
-// classification ip.To4() != nil.
-func ValidateIPv4CIDR(raw string, _ *Config) error {
-	ip, err := parseCIDRStrict(raw, "10.0.1.10/24")
-	if err != nil {
-		return err
-	}
-	if ip.To4() == nil {
-		return fmt.Errorf("not an IPv4 address (got %q; IPv6 addresses belong under family inet6)", raw)
-	}
-	return nil
-}
-
-// ValidateIPv6CIDR accepts an IPv6 address with an explicit prefix
-// length (e.g. 2001:db8::1/64). See ValidateIPv4CIDR for why the prefix
-// is required; the family rule mirrors the runtime's ip.To4() == nil
-// classification, so 4-in-6 forms like ::ffff:10.0.1.1/96 — which the
-// runtime classifies as inet — are rejected under family inet6.
-func ValidateIPv6CIDR(raw string, _ *Config) error {
-	ip, err := parseCIDRStrict(raw, "2001:db8::1/64")
-	if err != nil {
-		return err
-	}
-	if ip.To4() != nil {
-		return fmt.Errorf("not an IPv6 address (got %q; IPv4 addresses belong under family inet)", raw)
-	}
-	return nil
-}
-
-// parseCIDRStrict is the shared require-a-prefix CIDR parse for the two
-// family validators. It upgrades the two common operator mistakes to
-// targeted messages: a bare IP (missing /prefix-length) and outright
-// garbage.
-func parseCIDRStrict(raw, example string) (net.IP, error) {
-	trimmed := strings.TrimSpace(raw)
-	if trimmed == "" {
-		return nil, fmt.Errorf("missing value (expected address/prefix-length, e.g. %s)", example)
-	}
-	ip, _, err := net.ParseCIDR(trimmed)
-	if err != nil {
-		if net.ParseIP(trimmed) != nil {
-			return nil, fmt.Errorf("missing /prefix-length (got %q; the runtime silently skips addresses without one — write e.g. %s)", raw, example)
-		}
-		return nil, fmt.Errorf("not a valid address/prefix-length (expected e.g. %s): %v", example, err)
-	}
-	return ip, nil
-}
-
 // ValidateBGPHoldTime accepts a BGP hold-time in seconds: 0, or 3..65535.
 // FRR (and RFC 4271 §4.2 / §10) require the hold-time to be either 0
 // (hold timer disabled) or at least 3 seconds — the keepalive interval is
